various improvement to build, version bump for further updater testing
