+++ GitVersion.yml
@@ -1,5 +1,5 @@
 mode: ContinuousDeployment
-next-version: 0.1.0
+next-version: 0.1.1
 branches: {}
 ignore:
   sha: []
